Update WDPM.py
The code has been rated at 9.72/10 by pylint. For the remaining warning. Some of them relavant to the code structure. The body of the "size class" is too big. I am not sure if we can just leave it as it is. Because re-write this part might need big modification to the code. For the others, I haven't got solutions. Sometimes after get rid of the pylint warning, the code cannot work properly. I will further look into it.